requirements.txt
fastapi
uuid
requests
python-dotenv   